Remove timezone support from the API
Remove timezone support from the API as agreed during the last F2F meeting.
This makes the API much simpler, gets rid of the confusing TimezoneDirective
argument and avoids a lot of complicated cases related to timezone and DST.

We also stop using a Date to schedule an alarm. We now use a DOMTimeStamp,
which is a number of milliseconds past the epoch (e.g. Date.now() + n).